Fix inconsistent type of readdir(plus) offsets
`u64` was used:

* in `raw::abi::fuse_read_in` (the source of values for `readdir` and
  `readdirplus` arguments)
* in `raw::abi::fuse_dirent` (what `DirectoryEntry` and
  `DirectoryEntryPlus` are converted to)
* as `readdirplus` argument

`i64` was used:

* as `readdir` argument
* in `raw::reply::DirectoryEntry` and `raw::reply::DirectoryEntryPlus`,

This fixes this inconsistency by making them all use `u64`.

+++ src/raw/session.rs
@@ -3036,7 +3036,7 @@ impl<FS: Filesystem + Send + Sync + 'static> Session<FS> {
             );
 
             let reply_readdir = match fs
-                .readdir(request, in_header.nodeid, read_in.fh, read_in.offset as i64)
+                .readdir(request, in_header.nodeid, read_in.fh, read_in.offset)
                 .await
             {
                 Err(err) => {
@@ -3078,7 +3078,7 @@ impl<FS: Filesystem + Send + Sync + 'static> Session<FS> {
 
                 let dir_entry = fuse_dirent {
                     ino: entry.inode,
-                    off: entry.offset as u64,
+                    off: entry.offset,
                     namelen: name.len() as u32,
                     // learn from fuse-rs and golang bazil.org fuse DirentType
                     r#type: mode_from_kind_and_perm(entry.kind, 0) >> 12,
